Add confirmation reminder banner to account
When a user creates an account there can be a window of 7 days where
they can see their account dashboard but may not have clicked the
confirmation link send in an email.

Similarly, if a user changes there email address, there is a window
where that change may not have been confirmed.

This adds a new banner component to the account page that reminds the
user they have an outstanding confirmaton and offers them a link to
re-send the confirmation if their last one has not arrived.

The banner will only appear on the main three pages:
- Your account,
- Manage your account,
- Security

The form that the banner links to will also attempt to help the user by
filling in either:
- The unconfirmed email address the user is attempting to change to, or
- It will fall back to their main stored email address.

Where we cannot find something reasonable to fill that form with it will
end up blank.

They can make changes to these before submitting.