Decorate OAuth cross-domain links & forms with cross-domain param
If the user is sent from another service to the accounts domain for
OAuth, there are three cases:

1. The user already has a session.

2. The user does not have a session, and signs up: in which case we
add the _ga param to the "return to service" link on the confirmation
page.

3. The user does not have a session, and signs in: in which case we
add the _ga param to the sign in form target, and then append it to
the redirect_path in the controller.

In all three cases, the user then ends up in the Doorkeeper
authorizations controller, which will do one of two things:

1. Automatically consent.

2. Display a consent page: in which case we add the _ga param to the
form targets.

At the point at which Doorkeeper redirects the user back to the
service, it preserves any _ga query param.